requirements.txt
asgiref==3.8.1
Django==5.1.2
django-solo==2.4.0
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
